+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+* Added missed call notifications
+* Updated Android SDK
+* Added localization files
+
 ## 0.0.1
 
 * Initial release